Move OCR extractors to common package
