+++ src/main/java/uk/gov/hmcts/reform/bulkscanccdeventhandler/common/utils/OcrFieldExtractor.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.bulkscanccdeventhandler.transformation.services.utils;
+package uk.gov.hmcts.reform.bulkscanccdeventhandler.common.utils;
 
 import uk.gov.hmcts.reform.bulkscanccdeventhandler.common.model.in.ExceptionRecord;
 import uk.gov.hmcts.reform.bulkscanccdeventhandler.ocrvalidation.model.in.OcrDataField;
